+++ pytsite/formatters/__init__.py
@@ -1,8 +1,8 @@
 """PytSite Formatter
 """
-from ._formatters import Formatter, Bool, Int, PositiveInt, AboveZeroInt, Float, PositiveFloat, Str, DateTime, JSON, \
-    JSONArrayToList, JSONArrayToTuple, JSONObjectToDict
-
 __author__ = 'Oleksandr Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
+
+from ._formatters import Formatter, Transform, Bool, Int, PositiveInt, AboveZeroInt, Float, PositiveFloat, Str, Enum, \
+    DateTime, JSON, JSONArrayToList, JSONArrayToTuple, JSONObjectToDict